% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einsum-package.R
\docType{package}
\name{einsum_package}
\alias{einsum_package}
\title{Einstein Summation}
\description{
Einstein summation is a concise mathematical notation that
implicitly sums over repeated indices of n-dimensional arrays. Many ordinary
matrix operations (e.g. transpose, matrix multiplication, scalar product, 'diag()', trace etc.)
can be written using Einstein notation. The notation is particularly convenient for
expressing operations on arrays with more than two dimensions because the
respective operators ('tensor products') might not have a standardized name.
}
